Added ViewModels and Views for MUPs.Edit and MUPs.Delete. Still wiring everthing up.
