+++ VSProjects/Geomapmaker/Views/MapUnitPolys/CreateMapUnitPoly.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace Geomapmaker.Views.MapUnitPolys
+{
+    /// <summary>
+    /// Interaction logic for CreateMapUnitPoly.xaml
+    /// </summary>
+    public partial class CreateMapUnitPoly : UserControl
+    {
+        public CreateMapUnitPoly()
+        {
+            InitializeComponent();
+        }
+    }
+}
