+++ VSProjects/Geomapmaker/Views/MapUnitPolys/DeleteMapUnitPoly.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace Geomapmaker.Views.MapUnitPolys
+{
+    /// <summary>
+    /// Interaction logic for DeleteMapUnitPoly.xaml
+    /// </summary>
+    public partial class DeleteMapUnitPoly : UserControl
+    {
+        public DeleteMapUnitPoly()
+        {
+            InitializeComponent();
+        }
+    }
+}
