+++ VSProjects/Geomapmaker/Views/MapUnitPolys/EditMapUnitPoly.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace Geomapmaker.Views.MapUnitPolys
+{
+    /// <summary>
+    /// Interaction logic for EditCreateMapUnitPoly.xaml
+    /// </summary>
+    public partial class EditMapUnitPoly : UserControl
+    {
+        public EditMapUnitPoly()
+        {
+            InitializeComponent();
+        }
+    }
+}
